Take values from ssm env variables when applying configuration
Environment variables with ssm_ prefix are assumed to contain
configuration values for the service's configuration template.

This mechanism allows secrets in SSM Parameter Store to be easily
injected into containers.

This is the same feature that was attempted in an earlier,
reverted commit, with handling for the non-zero return code that
we may get from grep.