Fix installing of ytl certificate in suorekisteri trust store
The general handling of trusted certificates in run.sh is also
removed as unnecessary. Suoritusrekisteri is the one and only
special case that needs something like this.

Parsing of trust store password is fixed so the suoritusrekisteri
case works again. This hasn't worked for a long time, but has been
failing unnoticed until the -e option in run.sh was fixed.